+++ banyand/metadata/schema/checker.go
@@ -82,6 +82,13 @@ var CheckerMap = map[Kind]EqualityChecker{
 			protocmp.IgnoreFields(&databasev1.Node{}, "created_at"),
 			protocmp.Transform())
 	},
+	KindProperty: func(a, b proto.Message) bool {
+		return cmp.Equal(a, b,
+			protocmp.IgnoreUnknown(),
+			protocmp.IgnoreFields(&databasev1.Property{}, "updated_at"),
+			protocmp.IgnoreFields(&commonv1.Metadata{}, "id", "create_revision", "mod_revision"),
+			protocmp.Transform())
+	},
 	KindMask: func(_, _ proto.Message) bool {
 		return false
 	},